pyproject.toml
[tool.poetry]
name = "welwick"
version = "0.0.0"  # DO NOT CHANGE -- set during build
description = "Prepare your fortune on demand."
authors = ["Paul Lockaby <paul@paullockaby.com>"]
packages = [{include = "welwick", from = "src"}]

[tool.poetry.scripts]
welwick = "welwick:main"

[tool.poetry.dependencies]
python = "^3.10"
mastodon-py = "^1.8.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.10.0"
pre-commit = "^3.0.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
